# openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenQuester API",
    "version": "0.18.0"
  },
  "security": [
    {
      "SessionAuth": []
    }
  ],
  "paths": {
...
    "/v1/packages": {
      "post": {
        "summary": "Get upload links for every file in `content.json`",
        "tags": ["Packages"],
        "requestBody": {
          "description": "Data for package creation",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageCreationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Returns object of key-values where key is filename, and value is file upload link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Package is corrupted!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageCorruptedResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all packages",
        "tags": ["Packages"],
        "parameters": [
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PackagesSortBy"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OrderDirection"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PaginationLimit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PaginationOffset"
            }
          },
          {
            "name": "title",
            "in": "query",
            "description": "Search by package title, not case sensitive",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PaginationTitle"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all packages data successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedPackages"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/users/": {
      "get": {
        "summary": "Get all users info",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "sortBy",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UsersSortBy"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OrderDirection"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PaginationLimit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PaginationOffset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all users data successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsers"
                }
              }
            }
          },
          "403": {
            "description": "Permission denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionDeniedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Users not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserNotFoundResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/oauth2": {
      "post": {
        "summary": "Validate oauth2 token",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputOauthLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseUser"
                }
              }
            }
          },
          "400": {
            "description": "User have error during logging",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/games": {
      "get": {
        "summary": "Get all games",
        "tags": ["Games"],
        "parameters": [
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/GamesSortBy"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OrderDirection"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PaginationLimit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PaginationOffset"
            }
          },
          {
            "name": "createdAtMin",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FilterCreatedAtMin"
            }
          },
          {
            "name": "createdAtMax",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FilterCreatedAtMax"
            }
          },
          {
            "name": "isPrivate",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FilterIsPrivate"
            }
          },
          {
            "name": "titlePrefix",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PaginationTitlePrefix"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get all games successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedGames"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new game",
        "tags": ["Games"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameCreateData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Game created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameListItem"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Session-based authentication using cookies."
      }
    },
    "schemas": {
      "PageInfo": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "examples": [1000]
          }
        }
      },
      "PackageCreationInput": {
        "type": "object",
        "properties": {
          "content": {
            "$ref": "#/components/schemas/PackageCreateInputData"
          }
        },
        "required": ["content"]
      },
      "ResponseUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "examples": [1]
          },
          "username": {
            "type": "string",
            "examples": ["John"]
          },
          "email": {
            "type": ["string", "null"],
            "examples": ["john.doe@gmail.com"]
          },
          "discordId": {
            "type": ["string", "null"],
            "examples": ["123123123123"]
          },
          "birthday": {
            "type": ["string", "null"],
            "format": "date",
            "examples": ["2024-01-01"]
          },
          "avatar": {
            "type": ["string", "null"],
            "description": "link on file GET",
            "examples": [
              "https://s3.example.com/openquester-sample-bucket/e/ec/ec65cd3756490723f8bcc3bc432b94a7"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "examples": ["2024-01-01T00:00:00Z"]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "examples": ["2024-01-01T00:00:00Z"]
          },
          "isDeleted": {
            "type": "boolean",
            "examples": [false]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "examples": [1]
                },
                "name": {
                  "type": "string",
                  "examples": ["get_all_users"]
                }
              }
            }
          }
        }
      },
      "InputOauthLogin": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "examples": ["HSXxyt6nrbW5Gtqw5BWfHPYk9lxDOj"]
          },
          "tokenSchema": {
            "type": ["string", "null"],
            "examples": ["Bearer"]
          },
          "oauthProvider": {
            "type": "string",
            "examples": ["discord"],
            "enum": ["discord"]
          }
        }
      },
...
      "GameListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "examples": ["AH2S"]
          },
          "createdBy": {
            "$ref": "#/components/schemas/ShortUserInfo"
          },
          "title": {
            "type": "string",
            "examples": ["Game Title"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "examples": ["2024-01-01T00:00:00Z"]
          },
          "ageRestriction": {
            "$ref": "#/components/schemas/AgeRestriction"
          },
          "isPrivate": {
            "type": "boolean",
            "examples": [false]
          },
          "currentRound": {
            "type": ["integer", "null"],
            "examples": [0]
          },
          "currentQuestion": {
            "type": ["integer", "null"],
            "description": "Index (or id) of the currently active question in round if any",
            "examples": [0]
          },
          "playedQuestions": {
            "type": ["integer", "null"],
            "examples": [0]
          },
          "players": {
            "type": "integer",
            "examples": [5]
          },
          "maxPlayers": {
            "type": "integer",
            "examples": [10]
          },
          "startedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "examples": ["2024-01-01T00:00:00Z"]
          },
          "finishedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "examples": ["2024-01-01T02:00:00Z"]
          },
          "package": {
            "$ref": "#/components/schemas/PackageItem"
          }
        }
      },
      "GameCreateData": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "examples": ["Game Title"]
          },
          "packageId": {
            "type": "integer",
            "minimum": 0,
            "examples": [1]
          },
          "isPrivate": {
            "type": "boolean",
            "examples": [false]
          },
          "ageRestriction": {
            "$ref": "#/components/schemas/AgeRestriction"
          },
          "maxPlayers": {
            "type": "integer",
            "minimum": 2,
            "maximum": 15,
            "examples": [10]
          }
        }
      },
      "ShortUserInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "examples": [1]
          },
          "username": {
            "type": "string",
            "examples": ["John"]
          }
        }
      },
      "PackageTag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "examples": [15]
          },
          "tag": {
            "type": "string",
            "description": "A single tag for the package",
            "examples": ["anime"]
          }
        },
        "required": ["tag"],
        "description": "Individual tag object"
      },
      "PaginatedPackages": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackageListItem"
            }
          },
          "pageInfo": { "$ref": "#/components/schemas/PageInfo" }
        }
      },
      "PackageListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "examples": [1]
          },
          "title": {
            "type": "string",
            "examples": ["Package Title"]
          },
          "description": {
            "type": ["string", "null"],
            "examples": [
              "Package description, should be longer than title, probably"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "examples": ["2024-01-01T00:00:00Z"]
          },
          "author": {
            "$ref": "#/components/schemas/ShortUserInfo"
          },
          "ageRestriction": {
            "$ref": "#/components/schemas/AgeRestriction",
            "description": "Package age restriction"
          },
          "language": {
            "type": ["string", "null"],
            "examples": ["en"]
          },
          "logo": {
            "oneOf": [
              { "$ref": "#/components/schemas/PackageLogoFileItem" },
              { "type": "null" }
            ],
            "description": "Logo file for the package"
          },
          "tags": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/PackageTag"
            },
            "description": "Tags for the package. Can be null or an array of tag objects",
            "examples": [
              [
                { "id": 18, "tag": "anime" },
                { "id": 19, "tag": "movie" },
                { "id": 20, "tag": "general" },
                { "id": 21, "tag": "games" }
              ],
              [
                { "id": 22, "tag": "trivia" },
                { "id": 23, "tag": "fun" }
              ],
              null
            ]
          }
        }
      },
...
      "PackageItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "examples": [1]
          },
          "title": {
            "type": "string",
            "examples": ["Package Title"]
          },
          "description": {
            "type": ["string", "null"],
            "examples": [
              "Package description, should be longer than title, probably"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "examples": ["2024-01-01T00:00:00Z"]
          },
          "author": {
            "$ref": "#/components/schemas/ShortUserInfo"
          },
          "ageRestriction": {
            "$ref": "#/components/schemas/AgeRestriction",
            "description": "Package age restriction"
          },
          "language": {
            "type": ["string", "null"],
            "examples": ["en"]
          },
          "logo": {
            "oneOf": [
              { "$ref": "#/components/schemas/PackageLogoFileItem" },
              { "type": "null" }
            ],
            "description": "Logo file for the package"
          },
          "roundsCount": {
            "type": "integer",
            "examples": [5]
          },
          "questionsCount": {
            "type": "integer",
            "examples": [183]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "examples": ["Anime", "General"]
            }
          }
        }
      },
      "PaginatedUsers": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResponseUser"
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfo"
          }
        }
      },
      "PaginatedGames": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GameListItem"
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfo"
          }
        }
      },
      "OrderDirection": {
        "type": "string",
        "enum": ["asc", "desc"],
        "examples": ["asc"],
        "default": "desc"
      },
      "PaginationLimit": {
        "type": "integer",
        "examples": [10],
        "minimum": 1,
        "maximum": 50,
        "default": 10
      },
      "PaginationOffset": {
        "type": "integer",
        "examples": [0],
        "minimum": 0
      },
      "FilterCreatedAtMin": {
        "type": "string",
        "format": "date-time",
        "examples": ["2024-01-01T00:00:00Z"]
      },
      "FilterCreatedAtMax": {
        "type": "string",
        "format": "date-time",
        "examples": ["2024-01-01T12:00:00Z"]
      },
      "FilterIsPrivate": {
        "type": "boolean",
        "examples": [false]
      },
      "PaginationTitlePrefix": {
        "type": "string",
        "description": "Search by title prefix (start)",
        "examples": ["Anime"]
      },
      "PaginationTitle": {
        "type": "string",
        "examples": ["Gays"]
      },
      "GamesSortBy": {
        "type": "string",
        "enum": [
          "id",
          "title",
          "createdAt",
          "createdBy",
          "maxPlayers",
          "players",
          "startedAt"
        ],
        "examples": ["createdAt"],
        "default": "createdAt"
      },
      "UsersSortBy": {
        "type": "string",
        "enum": ["id", "is_deleted", "created_at", "username", "email"],
        "examples": ["createdAt"],
        "default": "createdAt"
      },
      "PackagesSortBy": {
        "type": "string",
        "enum": ["id", "title", "created_at", "author"],
        "examples": ["created_at"],
        "default": "created_at"
      },
      "PackageUploadResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "examples": [1]
          },
          "uploadLinks": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "examples": [
              { "uniqueIdentifier": "https://example.com/upload/file.jpg" }
            ]
          }
        }
      },
...
      "PackageCorruptedResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "examples": ["Package is corrupted!"]
          }
        }
      },
      "BadRequestResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "examples": ["Bad request"]
          }
        }
      },
...
      "UserNotFoundResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "examples": ["User not found"]
          }
        }
      },
      "PermissionDeniedResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "examples": ["You don't have permission to perform this action"]
          }
        }
      },
      "GeneralErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "examples": ["Authentication is failed, please try again"]
          }
        }
      },
...
      "AgeRestriction": {
        "type": "string",
        "enum": ["A18", "A16", "A12", "NONE"],
        "default": "NONE",
        "description": "Age restriction",
        "examples": ["A18"]
      },
      "PackageFileType": {
        "type": ["string", "null"],
        "enum": ["video", "audio", "image"],
        "description": "Type of media file for package-related files"
      },
...
      "FileInput": {
        "type": "object",
        "properties": {
          "md5": {
            "type": "string",
            "description": "MD5 hash of the file",
            "examples": ["d41d8cd98f00b204e9800998ecf8427e"]
          },
          "type": {
            "$ref": "#/components/schemas/PackageFileType"
          }
        },
        "required": ["md5", "type"],
        "description": "Basic file metadata for upload, used in place of full File model for input"
      },
      "FileItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": ["integer", "null"],
            "examples": [727]
          },
          "md5": {
            "type": "string",
            "description": "MD5 hash of the file",
            "examples": ["d41d8cd98f00b204e9800998ecf8427e"]
          },
          "type": {
            "$ref": "#/components/schemas/PackageFileType"
          },
          "link": {
            "type": ["string", "null"],
            "description": "GET link for file",
            "examples": ["https://example.com/file123"]
          }
        },
        "required": ["md5"],
        "description": "Basic file metadata"
      },
      "PackageLogoFileInput": {
        "type": ["object", "null"],
        "properties": {
          "file": {
            "$ref": "#/components/schemas/FileInput"
          }
        },
        "description": "Logo file structure for package creation. Only images acceptable"
      },
      "PackageLogoFileItem": {
        "type": ["object", "null"],
        "properties": {
          "file": {
            "$ref": "#/components/schemas/FileItem"
          }
        },
        "required": ["file"],
        "description": "Logo file structure for package creation. Only images acceptable"
      },
      "PackageEntitiesOrder": {
        "type": "integer",
        "description": "Order of the round/theme/question file in package data, should be unique, starting from 0",
        "examples": [1, 2, 3]
      },
...
      "SimpleQuestion": {
        "allOf": [
          { "$ref": "#/components/schemas/PackageQuestion" },
          {
            "type": "object",
            "properties": {
              "type": { "type": "string", "enum": ["simple"] }
            },
            "description": "Simple question type"
          }
        ]
      },
      "StakeQuestion": {
        "allOf": [
          { "$ref": "#/components/schemas/PackageQuestion" },
          {
            "type": "object",
            "properties": {
              "type": { "type": "string", "enum": ["stake"] },
              "subType": {
                "type": "string",
                "enum": ["simple", "forEveryone"],
                "description": "Subtype of the stake question. simple is basically auction. forEveryone means everyone place some amount of money and write answer",
                "default": "simple",
                "examples": ["simple"]
              },
              "maxPrice": {
                "$ref": "#/components/schemas/QuestionMaxPrice"
              }
            },
            "required": ["subType", "maxPrice"]
          }
        ]
      },
...
      "PackageQuestionUnion": {
        "oneOf": [
          { "$ref": "#/components/schemas/SimpleQuestion" },
          { "$ref": "#/components/schemas/StakeQuestion" },
          { "$ref": "#/components/schemas/SecretQuestion" },
          { "$ref": "#/components/schemas/NoRiskQuestion" },
          { "$ref": "#/components/schemas/ChoiceQuestion" },
          { "$ref": "#/components/schemas/HiddenQuestion" }
        ],
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "simple": "#/components/schemas/SimpleQuestion",
            "stake": "#/components/schemas/StakeQuestion",
            "secret": "#/components/schemas/SecretQuestion",
            "noRisk": "#/components/schemas/NoRiskQuestion",
            "choice": "#/components/schemas/ChoiceQuestion",
            "hidden": "#/components/schemas/HiddenQuestion"
          }
        },
        "description": "Polymorphic question union for all question types"
      },
      "PackageTheme": {
        "type": "object",
        "properties": {
          "id": {
            "type": ["integer", "null"],
            "examples": [92]
          },
          "order": {
            "$ref": "#/components/schemas/PackageEntitiesOrder"
          },
          "name": {
            "type": "string",
            "description": "Name of the theme",
            "examples": ["Opening", "Ending", "Science"]
          },
          "description": {
            "type": ["string", "null"],
            "description": "Description of the theme",
            "examples": [
              "In this theme you have to guess anime by it's opening song"
            ]
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PackageQuestionUnion"
            },
            "description": "Questions in the theme"
          }
        },
        "required": ["name", "order", "questions"],
        "description": "Theme structure for package"
      },
      "PackageRound": {
        "type": "object",
        "properties": {
          "id": {
            "type": ["integer", "null"],
            "examples": [17]
          },
          "order": {
            "$ref": "#/components/schemas/PackageEntitiesOrder"
          },
          "name": {
            "type": "string",
            "description": "Name of the round",
            "examples": ["Round 1 - Beginners round", "Round of music"]
          },
          "description": {
            "type": ["string", "null"],
            "description": "Description of the round",
            "examples": ["This round contains simple questions"]
          },
          "type": {
            "$ref": "#/components/schemas/PackageRoundType"
          },
          "themes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PackageTheme" },
            "description": "Themes in the round"
          }
        },
        "required": ["name", "order", "type", "themes"],
        "description": "Round structure for package"
      },
      "PackageCreateInputData": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the package",
            "examples": ["Trivia Night", "Anime pack 2025"]
          },
          "description": {
            "type": ["string", "null"],
            "description": "Description of the package",
            "examples": ["This is package description"]
          },
          "language": {
            "type": ["string", "null"],
            "description": "Language of the package",
            "examples": ["en", "ua"]
          },
          "ageRestriction": {
            "$ref": "#/components/schemas/AgeRestriction",
            "description": "Age restriction"
          },
          "logo": {
            "oneOf": [
              { "$ref": "#/components/schemas/PackageLogoFileInput" },
              { "type": "null" }
            ],
            "description": "Logo file for the package"
          },
          "tags": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/PackageTag"
            },
            "description": "Tags for the package. Can be null or an array of tag objects",
            "examples": [
              [
                { "id": 18, "tag": "anime" },
                { "id": 19, "tag": "movie" },
                { "id": 20, "tag": "general" },
                { "id": 21, "tag": "games" }
              ],
              [
                { "id": 22, "tag": "trivia" },
                { "id": 23, "tag": "fun" }
              ],
              null
            ]
          },
          "rounds": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/PackageRound" },
            "description": "Rounds in the package"
          }
        },
        "required": ["title", "ageRestriction", "rounds"],
        "description": "Input for creating a new package; author is set via authentication"
      },
...
      "PackageRoundType": {
        "type": "string",
        "enum": ["simple", "final"],
        "description": "Type of package round",
        "examples": ["simple"]
      },
...
    }
  }
}